refactor: split api.ts and PostgresAdapter.ts into focused modules
- Split api.ts (1666 lines) into api/ directory with 5 files:
  maps.ts (525), aliases.ts (506), normalize.ts (131),
  group-api.ts (194), index.ts (286)
- Extract schema operations from PostgresAdapter.ts (1203 -> 606 lines)
  into schema-operations.ts (633 lines)
- Update import paths in codemode/index.ts, tools/codemode/index.ts,
  api.test.ts, and PostgresAdapter.test.ts
- All 3314 tests pass, lint and typecheck clean
