+++ src/codemode/__tests__/api.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { describe, it, expect, vi, beforeEach } from "vitest";
-import { createPgApi, PgApi } from "../api.js";
+import { createPgApi, PgApi } from "../api/index.js";
 import type { PostgresAdapter } from "../../adapters/postgresql/PostgresAdapter.js";
 import type { ToolDefinition, ToolGroup } from "../../types/index.js";
