+++ src/codemode/index.ts
@@ -31,4 +31,4 @@ export { CodeModeSandbox, SandboxPool } from "./sandbox.js";
 export { CodeModeSecurityManager } from "./security.js";
 
 // API
-export { PgApi, createPgApi } from "./api.js";
+export { PgApi, createPgApi } from "./api/index.js";
